requirements.txt
pytest
google-generativeai